Wrap subprocess in try/catch blocks
Wrap subprocesses in try/catch blocks in the utils.js module and log
errors when process fails to start. Before this change, the extension
will fail to load when the command fails to run, for example when the
`tokens` command is missing or not in the PATH.

Be sure to use the fully qualified path to run the commands to make it
more clear what is not running were there is an error. The `tokens`
command is installed to /usr/bin/tokens by the debian/ubuntu packaging,
so use that path to run the tokens command. This makes it more clear
what is missing if the command is not found.

Be sure to prefix the new log messages with [openafs] so they can be
found in the large number of journalctl message, since all of the
gnome-shell messages go to the same location.